git-ffintegrateto: Usability: Print a header with the current (superproject) repo name if submodules have been processed
The subdo invocations have their own headers showing the current submodule, but after we're back to the superproject the push query directly follows, which is confusing, as the user is mentally still with the submodules.
Set a flag if submodule operations have (successfully) happened, and print an extra header then, to set the context for the following operations.