git-diffY: Unified diff output with delta but output is not to terminal
If the output of the git process is not a tty then git just completely ignores its core.pager setting (and therefore delta); cp. https://github.com/dandavison/delta/issues/101#issuecomment-593214649.
Thus, we need to pipe through delta ourselves in that case. Keep the original case where we just reconfigure delta to show side-by-side as that's a simple exec and the more common case.