hubcheats-submodules: Differentiate between with or without API changes
If a change in one submodule affects others, it's better to create a separate superproject commit (and amend any still compatible changes that haven't yet been recorded in the superproject). Each submodule just has a list of changes, and without the superproject commits that tie them together into compatible changes, one is left to use the commit dates to obtain revisions that go together.