ENH: List pull request approvals and ask for confirmation before reintegrating
When submodule integration is pending, if in a submodule just show and ask for the it; else for the superproject and all affected submodules (as that's what'll be covered).
When superproject integration is done separately (i.e. the user did the submodule integration in the individual submodules or declined to also do the superproject reintegration as well), just show and ask the superproject's pull requests.