CHG: Reuse the affected submodules for submodule iteration by default, allow reintegration of all submodules on the same branch via --same-branch
The submodule selection so far has been inconsistent; for counting it used the actual modified submodules on the current branch to master, but it then used subsamebr do iterate. In my workflow this distinction is not important, as only actual modified submodules should be on the same branch. Well, it can matter if the command is invoked multiple times in each submodule, as previous (now integrated) submodules are not on the branch any longer but on master. (Gosh, have we really evaded a bug here?)