git-cheat: ENH: Don't render span-level tags inside code sections (`...`)
With sed, that's a bit tricky: First put each code section on a separate line, with a leading ^G marker indicating that the line is a code fragment that still needs to be rendered and re-joined. With code sections and anything else neatly segregated onto separate lines, a second sed stage can rejoin the separate things, collecting everything in the hold space first.
We go with the newline separation of the H command and just do the join at the end, where duplicate newlines indicate actual newlines and single newlines are just artifacts of the hold space pushes.