FIX: selected-command with a command that expects addressing from the working copy root need to be invoked from there
Most sources of filespec (diff, showfiles etc.) printf filespecs relative to the working copy root, but a Git command that takes filespecs interprets those relative to the CWD. When the command is executed from a subdirectory, that doesn't match, and I get this:

> fatal: ambiguous argument 'tests/actions.sh': unknown revision or path not in the working tree.
> Use '--' to separate paths from revisions, like this:
> 'git <command> [<revision>...] -- [<file>...]'

I rarely operate in subdirs, that's why I did not recognize this for a long time.
git-selected-command works with arbitrary selections, instead of modifying it to "chdir $(git root)" (or adding such flag), it's better to compose this functionality and wrap the affected commands with my newly extended git-root.
Some callers operate on branches, others have e.g. git ls-files (which prints filespecs relative to the CWD); these must not use git-root (and have gotten comments about this).