ENH: Also store and compare keys without the author timestamp to recognized reworked commits
A rebase does not change the author date, but if the same change is recommitted, the date is different.
Move the timestamp to the front so that it can be easily cut off and store the resulting log without the timestamp in a separate incomingReworkedKeys dict.
Duplicate the rebase detection and reporting into reworks (rebase are "higher quality" and therefore have precedence).