hub-subpull-request: CHG: Use the log from the submodule also when executed in superproject
To have a consistent result regardless of where this is invoked. The previous manual invocation couldn't easily do this, but the program can.
For the implementation, we now invoke ourselves in the submodule iteration, but pass the superproject message to avoid querying that again for each submodule. (We could do the same for git-servername, but as that one is derived from the branch (which is the same everywhere), we don't need to go to the superproject for that.)