hub-subpull-request: ENH: Obtain the pull requests from other submodules and add those for reference
With hub-supersubpull-request, all submodule pull requests will be linked via the superproject's pull request. But here, there will only be pull requests in the submodules, so no linking so far. As pull requests for submodules are created, it would be nice if each maintains a list of previous ones (the only flaw is that when executed in the superproject, the order of pull requests cannot be influenced so far).
We can do that by obtaining the existing pull requests in other submodules from git-subchanges --only-pull-requests. As without a <range> (that we don't have, this is about same-name branches), we have to pass the list of affected submodules, we invoke ourselves recursively with --existing-pull-requests, determine the submodules and then their potential pull request titles. If there's such output, we splice it into the template for the current submodule's pull request, then repeat the same for the following pull request (which will find the previous pull requests as well).