git-ffintegrateto: Add a comment with a fast-forward integration note to the merge message template
In my Vim integration, the :GitcommitMergeL* commands cannot grab the commits, because the situation is different: I'm not on the target branch, but instead the target branch has been merged into the current one. However, the synthesized merge message template reads "Merge branch CURRENT" instead of (the technically correct, but misleading on the intention) "Merge branch TARGET into CURRENT". Because of the other branch is implicit, I also cannot simply turn around the range, and would have to assume the master branch is targeted. Even worse, my local master may be out-of-date, and instead the tracked master may be used!
Let's solve this with a comment that helps Vim detect this situation and obtain the target branch.