CHG: Use highest version number tag instead of newest for uptodate, and make that the (configurable) default in git-up, too
Git can only do sorting by creator date if the commits have been fetched. For git-up, that's no problem; the tag forwarding happens after fetching. But for git-uptodate, this isn't feasible, as fetching first would contradict the intention of a quick and lightweight check.
In reality, the distinction mostly shouldn't matter, as forwardable tags likely adhere to a versioning scheme, anyway, and versions are typically released in increasing order.