git-ffintegrateto: Also try to obtain the pull request based on the tracked branch's commit
When the superproject branch consists of multiple commits; i.e. we'd like to maintain the history of how the feature grew, a submodule reintegration is added as an additional commit on top, but that isn't pushed, so the pull request still is about the tracked remote branch, not the progressed local branch, and previously that pull request then wasn't found.
Introduce suffixArguments and add both HEAD and $trackedBranch for consideration.