Add git-previouscommit-command and make br*rb[[n]i] commands select a commit from the range
git-bmrbi rebases on master, not on the range that bmlg covers; that's confusing. I'll instead want to use mrbi for such an unconditional rebase, and instead align the bmrbi variant. As related commands like bmwipe also prompt for the selection of a commit from the range; this makes sense for rebasing as well, and elegantly avoids the problem that we'd have to resolve the base commit.
However, there's still one complication: Rebase needs the predecessor of the selected commit; let's add git-previouscommit-command for that to handle this generically.