git-ffintegrateto: BUG: Also try to obtain the pull request based on the tracked source branch's commit (after 7b6f093)
When the superproject branch consists of multiple commits; i.e. we'd like to maintain the history of how the feature grew, a submodule reintegration is added as an additional commit on top, but that isn't pushed, so the pull request still is about the tracked remote branch, not the progressed local branch, and previously that pull request then wasn't found.
Introduce suffixArguments and add both HEAD and $sourceBranch for consideration.
My previous fix attempt mistakenly used $targetBranch and $targetTracks, but that's master / origin/master, not the feature branch that we're on.

(This corrects commit 7b6f093d4af3e8a2d84a54293b3a7d1f8ecdcfb6)