git-lch: BUG: Hunks for all modified files are listed
We've only used a passed list of <path>s to filter the commits that modified those files, but then showed hunks for all files in those commits.
git-showh now supports filtering of files; extract the files and pass them along. To integrate that into xargs (but keeping the argument batching), we need placeholderArguments to move the <commmit>s (which xargs can only append) into the right position before the <path>s. As an optimization, only do this if necessary.