git-reintegrate: ENH: Offer to pull if a GitHub rebase is done
When used standalone, the recommendation to just exit is fine, as only a pull and potential branch cleanup is necessary after that. But if reintegrate is part of a larger workflow, the abort means additional manual work. It would be much more comfortable if the script could handle that.
Add a separate [P]ull choice that pulls upstream changes (and ensures something got indeed updated). Local checks are still run (after the fact), and postprocessing steps like branch deletion still run, too.
Include the PR URL in the message so that it can be easily launched - we need to visit GitHub to perform the manual rebase there.