hub-subpull-request: ENH: Include any changes in other submodules, not just existing pull requests
When submitting a PR for just one submodule, but (an)other submodule(s) has only small adaptations that don't warrant a separate PR, it's useful to reference those commits as well. (The whole list is just an offer and can be edited out.)
git-subchanges now has a handy --exclude-submodule SUBMODULE, so we don't need to filter out our own (which is obtained via git-supersubPath, and passed into the recursive call).
We now need a range for the commit list, so git-subchanges becomes hub-bomsubchanges.