hub-push-this-branch: Exit 99 if on detached working copy
When the submodules have been integrated and pushed, and even the superproject got merged and only the final push to master fails (e.g. because the GitHub action failed or other commits have sneaked in in front of us), a restart of the integration fails because the submodules cannot be pushed. (And hub-superonlyreintegratetom also cannot be used because it detects submodule changes.)
Make hub-push-this-branch detect a detached working copy and return 99 then; git-ffintegrateto will correctly interpret this as no-op and continue then.