hub-delete-merged-branches: Exit 99 if remote branch got auto-deleted and local branch got deleted, too
When the submodules have been integrated and pushed, and even the superproject got merged and only the final push to master fails (e.g. because the GitHub action failed or other commits have sneaked in in front of us), a restart of the integration fails because the submodule branches cannot be deleted. (And hub-superonlyreintegratetom also cannot be used because it detects submodule changes.)
Make hub-delete-merged-branches check for the local branch existence and return 99 then; git-ffintegrateto will correctly interpret this as no-op and continue then.