git-up: BUG: Remote gets re-fetched even though it had just been fetched
I noticed this when my push to main failed due to concurrent commits and I had to restart the reintegration. It would still mention the same outdated last fetch date.
The reason is that "hub supersubreintegratetom" sets HUB=t, which my git-wrapper interprets as choosing the hub-fetch variant over git-fetch, and that one is only meant for the fetch of remotes of contributors, not for the ordinary fetch, and so it doesn't update the fetch date.
Explicitly invoke git-fetch here, bypassing the usual wrapper (which abstracts whether the built-in command actually got extended or not).