ENH: Add [uo]detach{,[recent]sbr[c]} variants
To check out a remote branch in "read-only" mode, without the intention of making modifications to it. I could just use [uo]co for that, but then I have to remember to clean up the local branch, which is a hassle.
git-checkout-remote-branch needs special handling for --detach, as it's incompatible with branch creation through -b|-B, resulting in "fatal: '--detach' cannot be used with '-b/-B/--orphan'"