Add reauthorasme alias
This is helpful after realizing that a new repository hasn't gotten my custom user configuration yet and I've done [several] commit[s] with my private Git account. (For a sequence of commits use interactive rebase.)
Note: If just the email address is wrong, the default git-mine-command won't do anything, as the username is identical. We need to pass the full "username <email>" for this to work.