ENH: Add logmailmapped alias
To find commits where a differing author has been used that is mapped to the currently shown value.

Note that we have to make up or own "even more fuller" log format, as no built-in one has both original and mailmapped fields. Git itself doesn't use the format-string for its built-in formats; those actually are hard-coded in the C sources and that's how they get the 4-space indented subject and body (and omitted empty lines if there's no body). We can only weakly emulate that; subsequent body lines won't have the indent and we have additional empty lines if there's no body.