git-rblastfixup: BUG: A passed <revision-range> makes the rebase a no-op
Because that range is added to the ${lastCommit}^ that limits the original commit search to begin before the found fixup commit. With that, the search will cover the entire range, and find that fixup commit itself again!
We need to parse a passed range separately (into $rangeArg), and additionally store the start of the range as $rangeStart, as that's what needs to be prepended to ${lastCommit}^.
The range isn't part of $gitArgs any longer, and only is explicitly passed to the fixup search command now.