gitconfig: Exclude archived branches and commits by default
So that git-up ignores the archive tags in its search for the tag with the highest version number to check out.
I've had LTS to revert to very old versions instead of tracking main, and the reason was first a stray tag and then the archived branches.
Multi-valued keys are automatically joined from all scopes, so we don't prevent repo-local additions by adding a user scope, and neither does a local addition eclipse the user-scoped value.