ENH: Add s{,c,w,a}{,st,i,-I,last} variants of ss{,c,w,a}{,st,i,-I,last} that don't query
But instead directly take the last commit of the range. This likely doesn't make much sense for *{x,q,z}, but let's have them as well for consistency.
For the basic "s" (show last commit in passed range), the existing "short status" alias has to go, but I haven't used that at all (just "git st").
sg is the same as wherelastshowg, but implemented differently, i.e. directly picking the last commit instead of reversing the log and picking the first.
The *last{st,i,-i,last} aliases are equivalent to *1 (well, they actually showed the full commit, not just the diff for the modified / staged / etc. files, but that was more of an implementation detail rather than an intentional decision, and I'm glad this got ironed out) and therefore redundant.