ENH: Allow further restrict ref names via -e|--exclude and -i|--include
Aliases like lgtagged and lgupstream already constrain the refs via --decorate-refs=<pattern>, so this isn't available to the user any longer; --decorate-refs-exclude could be used, but would be cumbersome, as the alias's filter would have to be repeated.
Let's add a separate way of specifying both positive and negative patterns on top of the alias's.