git-super{only,sub}reintegratetom: ENH: Also check for open PR comments and confirm
It may just be courtesy to resolve PR comments prior to merging, or it can actually be made mandatory via a branch protection rule ("Require conversation resolution before merging", required_conversation_resolution=True). We could check to give a better prompt, but as the target branch is involved and the setting would have to be queried, let's assume that the user remembers which rules the repo operates under and just ask generically.