hub-pr{approval,comments}: Robustness: Split apiCall from rendering and handle unknown PR ID
If the supplied PR ID does not exist, a different response is returned:
> {"message":"Not Found","documentation_url":"https://docs.github.com/rest/reference/pulls#list-reviews-for-a-pull-request"}

jq's parsing barfs on that, which isn't nice:
> jq: error (at <stdin>:0): Cannot index string with string "body"

Apply the error handling from hub-branch-protection by splitting the API call and the parsing via jq into two separate functions, and only go this route if the API call was successful. In case of an error, extract the error message (via a separate jq call), print that, and fail.