git-rev-range-this-branch: Optimization: "git brbase" takes many seconds to fail in a repo with many tags
Because we need to iterate through all tagged commits until we finally realize that there's no matching branch.
"git describe" unfortunately doesn't offer an exclusion for tags or remote branches when --all is used, but we can turn this around and only include all local branches. This means that far fewer commits are iterated on before failing.