git-loglimit-to-range-command: BUG: Range that includes the first commit leads to error in git-showfiles
> fatal: ambiguous argument '8634ad8f00213df830a67fa0141723959bf3bdc4^..8980cf7340321317e69ed933913754e0f715cc03': unknown revision or path not in the working tree.

Instead of just using REV^, we have to check whether that actually exists. Unfortunately, the correct inclusion of the very first commit depends on whether a diff or log is used; for the latter, we can just omit the (end) range, but for a diff / show we need to specify the full range.