git-subdo-core: ENH: Store previous submodule environments and recall via --repeat
This is useful when the iteration was based on a predicate (e.g. dirty submodules), and now the dirtiness has been changed. (For example, adding, committing, and then pushing submodules. The first two are still covered by the dirty predicate, but after committing, the dirtiness is gone; switching to a different predicate (wips) is difficult and requires thought. One could have first just printed the dirty submodules and then later used --for ... ;, but that's cumbersome to do.)