git-wcdo-core: ENH: Store previous working copies and recall via --repeat
This is useful when the iteration was based on a predicate (e.g. dirty working copies), and now the dirtiness has been changed. (For example, adding, committing, and then pushing working copies. The first two are still covered by the dirty predicate, but after committing, the dirtiness is gone; switching to a different predicate (wips) is difficult and requires thought. One could have first just printed the dirty working copies and then later passed them to (unwrapped) git-wcdo, but that's very cumbersome to do.)