Adapt: Make custom subcommand specializations of subdo pass a custom name
When aliases or subcommands reuse subdo (interestingly, I have plenty such for submodule iteration, but no generic specializations for wcdo et al.), this has the risk of that run overriding the store that a user currently has in progress, and whose repeat would then use the wrong submodules, which would be a grave problem (as it's hard to detect). Internal reuses of subdo should be segregated into their own stores. As the majority are aliases, and shell variables are hard to pass there, let's introduce an internal --subdo-command-name COMMAND-NAME parameter instead, which is easy to add to all aliases.