FIX: Need to pass reintegrateArgs to ourselves on continuation and parse our own --ff-target-to-tracked, too
When continuing, any arguments for git-reintegrateto passed to us need to be forwarded as well.
When the target branch is not up-to-date with the tracked remote, we add --ff-target-to-tracked ourselves, and that should be remembered for a continuation as well (although the only negative effect of not doing so would be a repeated user query).