BUG: Submodules are derived from the local target branch when it's behind the tracked remote
But the actual integration happens with the tracked remote branch (via $targetIntegrationSourceBranch). So if any others' commits that have not yet been added to the local target branch have modified submodules, these are reported by showsubmodules now as well, and pushing / deleting is attempted (but won't work, as there's no integration branch in these submodule repos).
The targetIntegrationSourceBranch needs to be used not just for merging / rebasing, but for the following submodule pushes and deletions as well. This was an oversight when this remote-branch forwarding was implemented.