Pass the targetIntegrationSourceBranch to the command continuation
The fact that the local branch is behind the tracked remote and the user wants to integrate to that gets lost if reintegration is aborted.
The continuation could re-detect the situation and ask again, but we can actually use the --ff-target-to-tracked git-reintegrateto command-line option to detect that outcome.
Parse that separately and set a new isFastForwardTargetToTracked flag, which is then used in the continuation branch to re-initialize the targetIntegrationSourceBranch accordingly.