CHG: DWIM: Default to cloning the current branch instead of reusing the original branch name
So we're not forced to provide a branch name for all but one bifurcation. It's also more save this way as the original branch does not get changed, and a redo of the operation if possible (even without removing the bifurcations, as the cloning will choose subsequent numbers on the next attempt).