BUG: ffintegrateto adds "# via fast-forward integration to main" into the actual commit message
When a merge is desired but there have been no changes on the destination branch. (That made it so hard to track down!)
In that case, git-mergeancestor is invoked and that one doesn't do the stripping that "git merge" does automatically. Unfortunately, "git committree" does not understand --cleanup=strip, so we have to reimplement that functionality ourselves (in git-commitMessageStrip).