hub-integrate: Check whether the workflow build actually succeeded and else ask for retry
I've had the branch build fail, and the script just was happy that the workflow had concluded and already removed my local branch and waited for the remote branch to vanish.
Issue another (silent) workflow query to ensure that it was successful. If not, offer to launch the GitHub run (as restarting the job most likely will be attempted), or just wait for the workflow to be restarted (if the user has the page already open).
"Proceed as if it had succeeded" would act like before (NOT recommended), and "Quit" just exits.