git-stash-rename: FIX: Use the previously renamed subject for renaming
This is obtained via --pretty=format:%gs (reflog subject), but it cannot be obtained via "git log", unfortunately. Neither does "git stash show --no-patch" work. So we can only use "git stash list", but it does strange things (in Git version 2.41.0) when passed a stash@{1}: It prints the whole list _and_ all stashes again starting with the passed one. sed to the rescue; we also need it to cut away the reflog prefix.