Refactoring: Introduce GIT_UP_REDIRECTION and replace the 2>&1 | git-hushup with it
git-hushup actually only needs the combined output and errors from the update (i.e. fetch / rebase / pull) operation itself; it's not interested in any delta logging [to the terminal] or the new after-update check output. In fact, that output (when it comes after submodule updates) confuses the parser!
Introduce a gitUpdateOperation() wrapper that applies the new GIT_UP_REDIRECTION setting only to the update operation itself, and modify all git-up clients that add the hushup to use that.