Consistency: CHG: lghi[ghlight] now takes text|/regexp/ like lgg and as the last argument, too
Aliases like mlghi and tdlghi were completely broken, as they did not (and actually could not) consider that the first argument (which in those cases was part of the alias itself) is the PATTERN. (I've made some adaptations like passing -3 instead of -2, but that also doesn't help that the order is wrong.)
We need to take the pattern as the last argument, just like lgg does.
Add an alternative --grep <pattern> variant here, too, and interpret a non-delimited argument as literal text.