ENH: br...lghi{,st,i,-i} use greyed-out one-more log and also highlight in there
Having the one-more log in normal color output is inconsistent with e.g br...lg. Unfortunately, we cannot simply use the --one-more-command greyonelinelog, as that would receive the additional text|/regexp/ argument as well.
But if we use a grey variant of lghighlight, that'll do - it'll even highlight in the greyed-out one-more line!
Unfortunately, we need grey variants not just for lghighlight, but also for lghi{st,i,-i}, so let's build a generic greyonelineloghighlight wrapper that takes the actual command as its first argument (which git-rev-range now can handle through use of eval()).