FIX: br...lgby doesn't work with --one-more-output (any longer)
Because git-rev-range now invokes the GIT-COMMAND (so others-command ...) for that additional command now.
Using an explicit --one-more-command greyonelinelog (which would be even better, with the greying-out) also fails due to the AUTHORS being passed to it - a greyothers-command that swallows that would be needed.
Using --one-more-command ^ also doesn't help; it just passed one additional revision to others-command.
It's not so important here, and br...logby never had this, neither, so just drop it.