hub-integrate: Take (and ignore) --no-checks|-c|--command|--exec commands
When git-super[sub]reintegratetom is invoked with a custom check argument (e.g. --no-checks), hub-integrate complains about that. Splitting off all of these various arguments in two scripts is tedious; also, we might support local checking after-the-fact as a feature in hub-integrate later.
Therefore, accept these (undocumented) arguments and ignore them. Print a note so the user is informed of that, too.