git-diffselected: ENH: Allow passing a range as interpreted by git rev-range
The semantics of r1..r2 and r1...r2 are swapped between rev-range and diff. Normally I'd use "git-rev-range --with-range ...", but for [dt]tds, I need the RANGE both to be passed to showfiles _and_ diffselected, but I cannot pass different formats to both, and the lookup has to happen first.
Pragmatic solution is to teach diffselected adapt the range, just like we're already doing in git-showfiles.