Add *tds variants for {in,out} and iterating over branches
For the latter, first the union of all changed files is built (via GIT_SELECTED_COMMAND_DEFAULT_FILES), the user is prompted for the selection, and the result is passed to the iterating branch diff command.
So, there's two iterations: First to obtain all changed files, then to do the diff branch-by-branch.