Robustness: Just add updated submodule(s), not just any superproject changes
"git amenu" and "git cu" are a bit dangerous in that they also pick up any changes in the superproject, not just the updated submodule references. This should not happen, but better be safe and explicitly include on the affected submodules - we have that list, anyway.
(There's still a small risk of manually staged files - but who would manually manipulate the index during a reintegration?)