ENH: Detect modified submodule references in superproject and offer to commit them
When a submodule merge fails (e.g. due to conflicts), and that submodule's reintegration is continued (hub reintegrate --continue), the amend / commit of the changed submodule reference (what's now been extracted into completeSubmoduleReintegrationInSuperproject()) does not happen.
Fortunately, there's no need to add a git-supersubreintegratetom --continue, as we can detect that situation (affected submodules are dirty, but submodules themselves are done) and (after confirmation) can do this and continue with the regular flow.
Note: Other uncommitted changes in the superproject would be ignored for that, but when the superproject itself is reintegrated the working copy then has to be totally clean.