ENH: Add git-onemorelg and use that in git-selectedcommit-command --one-more
A plain A..B range can be extended by appending "^": A^..B, but even that requires parsing. Not all ranges can be extended. It's worse if FILE(s) are involved. A range just selects commits, regardless of which files are affected.
For commands like catselected and cpselected, we'd like to include one more commit _that affects the passed FILEs_.
To obtain that, let's parse off the range (implementation taken from git-selectedcommit-command), and execute once with that range, then without any range, and include the line just beyond the first log's output. This means that it only works with one-line output (and no coloring), but for the internal use this is fine.