Add diffbin
We can define a diff driver for binary files, but that is bound to a particular filetype, mapped via gitattributes.
,----[ gitconfig ]----
| [diff "binary"]
| 	binary = true
| 	textconv = xxd
`----
Unfortunately, it's not possible to map all binary files, nor is it possible to enable this on-demand. (https://stackoverflow.com/a/67841092/813602 describes a hack that uses "cat" by default (inefficient), and even that won't work if there's a particular diff driver already configured for a file pattern.)

Instead, extend my custom difftool definitions, making this a bit more effort (as we need to handle color and delta pager ourselves), but fully orthogonal to the diff drivers.