hub-integrate: Check for draft PR and offer to make it ready
But only if --offer is given (and we're asking the user, anyway). In normal use, a draft PR is an error that aborts the command. The user is supposed to make the PR ready on their own (which is easy via "gh pr ready ID").
For our supersubreintegrateto workflow, the offer now correctly handles a draft PR in the superproject (because submodules need to be merged first) - however, remote merging in GitHub is only supported if those submodule reintegrations all have been fast-forwards that don't require a superproject update (as that update happens locally).
Note: We didn't even have to use hub-pr-isready; the isDraft flag can be added to the existing query (but having that separate command is still good for manual use).