git-reattach: BUG: git-checkoutbranch resets the branch
Which is **not** what we want here; rather, we want to checkout the found branch as-is.
Unfortunately, git-checkoutbranch is explicitly made for the other way around (checkout out a branch based on the current or passed revision); we have to use the built-in git-checkout (or rather our wrapper with the additional submodule handling).
This means that we have to duplicate the git-logdelta handling that git-checkoutbranch provides; writing an explicit wrapper for that is not worth it, and I'm also reluctant to build that into the git-checkout wrapper itself.