FIX: Remove {,u,o}[mn]bifurcate variants (after 6ed18a0)
Redistribution of commits onto different branches only makes sense for a commit range starting from the current commit, so all br... variants and the bifurateselected ones.
In contrast, mrb et al. are used for rebasing a difference of commits onto the start point of that difference; i.e. catching up to upstream changes. Though this could theoretically be combined with bifuration, that would be complex and weird.
Put another way, mrb is for catch-up of the feature branch to master, and bmrbi for rebasing all feature commits (e.g. to apply reorders, fixups, and squashes). Bifurcation purely is a more elaborate variant of the latter.

(This corrects commit 6ed18a0c452a30ce5c2985d5172d10f5d5954d82)