FIX: Remove {,u,o}[mn]rb[n]i aliases
Rebasing onto a target mostly only makes sense with a plain "rebase". It could be combined with interactiveness, but editing is not on one's mind when updating the branch with upstream changes. (The only think I can think of is knowing that a local commit conflicts with an upstream change, and one wants to drop it. But that could also be done prior to the rebase (bmrbi followed by mrb), or one explicitly passes the flag (mrb -i).
Getting rid of these variants allows me to align the rebase aliases with the br...rb variants; there, the non-interactive rebase doesn't make sense (because it	iterates over the current range without touching fixups), and I can redefine br...rb to rebase to the base commit (for some variants like tobrb, stablerb (but not bmrb which is duplicate of mrb)).