CHG: Redefine br...rb to rebase on top of the base revision / remove more variants where this is a no-op
It's not just the timespan variants where a plain rebase is a no-op (as no fixup are touched and rb[n]i has to be used). Other variants, especially all "revisions to branch", have that behavior as well.
Only tobrb, pcrb, stablerb, b[2][lp]rb refer to a revision that (can) be a non-predecessor of the current commit.
Redefine those to work like {,u,o}[mn]rb; i.e. rebasing on top of it. Where there's an existing alias, the alternative is omitted (e.g. no bmrb because mrb is already there).