ENH: Add [uo]d[hi] aliases
For diffing of the last commit / index to the remote; the diff needs to be reversed as the rev-remote-command appends the remote but we want this to be the first argument.
This is useful to see the changes of a rebased local branch. "git od" only works if the working copy and index are clean; "git ooutd" would show the entire diff, not considering what's already in the rebased remote commit.