git-reintegrate: Check for target being simply ahead of tracked
When there are local changes not yet pushed to the remote, we get this confusing question for merging:
> Your master branch is ahead 1 vs. origin; do you want to stop here and merge that first?

Confusing because this is simply resolved by a push; no merge required. We've so far just checked for a fast-forward of the local target to the remote, but didn't check the other way.
Add a isAheadOfTracked() function that simply prints a note.