git-difflines: USE $BASE (set by "git difftool") instead of assuming the last filespec exists
As that's not the case if two <commit> <commit> are passed, or even if the second <commit> is HEAD but the working copy is dirty. Then "git difftool" will need to use a temp file, which does not correspond to a file under version control and therefore causes an error when passed to "git blame".
$BASE will always point to a versioned file (or so I hope).