Add git-selectedannotatediffcommit-command and {correct,fix{up,amend,wording}[rb],commit{identical,like,relate},amendrelate}ad{st,i,I} aliases
Filtering commits by the same files is nice, but the killer is correlating the changed lines to a previous change in that line.
Note: I think the *i aliases (targeting the staged and about to be committed changes) are most useful, but I've added the working copy (*I) and all modification (*st) variants as well for consistency.