FIX: Branches passed to *brc[do] are processed in passed order, not in creation order
Because --branches BRANCH [...] ; are not passed to the *_SOURCE_COMMAND. However, if that is git-localbrc, we do want to pass them to get them into creation order. (Alternatively, each *brc wrapper would have to forgo the *_SOURCE_COMMAND and again invoke this on its own, but also parse out the --include-* flags - very ugly.)
Introduce a *_SOURCE_SORTED flag that skips the direct use of the passed branches.