hubcheats-submodules-serial: Rebase only possible (without conflicts) when amending
If there's the history of commits in the superproject, but the submodules have been rebased, we cannot (easily) rewrite the superproject history; Git does not track submodule commits before and after the rebase there.
We have to keep the history as it is, referencing the original commits. Only the housekeeping commit can be amended and kept up-to-date.

"git subsamebr *rb" needs -i because Git does terminal updates during the rebase, and the output can be long due to the "warning: skipped previously applied commit".