hub-[super]subpull-request: BUG: "ERROR: No submodule header found." when specifying another branch
"git subsamebrdo" is used to iterate over submodules, but that does not consider if another branch is the target; we need to iterate _those_ submodules, not the ones on the current branch! As I often work on the same submodule(s), this so far didn't occur.
Pass the new --branch parameter if --head is given.