hub-su{b,per,persub}pull-request: Robustness: Only pass --head if not pointing to HEAD
"hub pull-request" defaults to HEAD and --head HEAD is superfluous. So far this didn't matter at all, but with the "git subsamebrdo --branch", this does not ensure that it's the _current_ branch in the submodule, only that such a branch exist. This is a highly theoretical setup, but better be safe.