ENH: Detect rebases and skip those logs
If the submodule and superproject have been rebased on top of e.g. master, the log looks like this:
,----[ log ]----
| Submodule XxXxXx 58a38bc..dfaf487:
|   > XXXX~0000~Xxxxxxxxx~~Xxxxxxxx~xx~xxxxxxxxxxx~xxxxxxxxxxx
|   > XXXX~0000~Xxxxxxxx~~Xxxx~~Xxxx
| Submodule XxXxXx 1c04f1c..cc7e0c0:
|   > XXXX~0000~Xxxxxxxxxxx~~Xxx~Xxxx~xxx~xxx~xxxxx~xxxxxx~xx~Xxxxxxx~xxxxXxxxxXxxx~~~~xxx
| Submodule XxXxXx 153580e...1c04f1c:
|   > XXXX~0000~Xxxxxxxxxxx~~Xxxxx~xx~xxxxxxx~XxXxXx~xxx~Xxxxxxxx~xxxxxxx
|   > XXXX~0000~Xxxxxxxxxx~~Xxxxxx~xxx~xxxxx~xxxxxx~xxxxxxxx
|   < Xxxxxxxxxxx~~Xxx~xxxxxx~Xxxx~xxxxxxxx~xxxxxxx~xx~x~~x~~x~~xxxxx~~xxxxxx~~xxxxxx~xxx~xxxxx~~xxx
|   < Xxxxxxxxxxx~~xxxxx~xxx~xxxx~xxx~xxx~XxXxxxx~x~Xxxx
|   > XXX~~Xxxxxxxx~xxx~xxxxxxx~xxxxxxxx~~xx~xxx~~xxx~~xx~xx~xxxxxx~~xxxxx~xx~xxxxxx~xxx
|   < XXX~~~xxxxx~xxxx~xxxx~xxxx~xxx~xxxx
|   < XXX~~~xxxxx~xxxx~xxxxxxxx~xxxxx~XxxxXxxxxxxXxxxxxxxx
| Submodule XxXxXx 414ab8d..3589a4c:
|   > XXX~~~xxxxx~xxxx~xxxx~xxxx~xxx~xxxx
|   > XXX~~~xxxxx~xxxx~xxxxxxxx~xxxxx~XxxxXxxxxxxXxxxxxxxx
`----
The rebase can be detected by the use of "..." instead of ".."; including all commits as-is is messy; we'd better skip the rebase itself (i.e. the part with the most turmoil), and only include the normal commit ranges (even if that means that earlier ones have commits since merged to master, and therefore not belonging here; but these can be detected or even edited out later).
Now, there can be multiple "Submodule FOO" seconds; we need to adapt the processing for that, too.