hubcheats-submodules-serial: BUG: superproject rebase leads to conflicts and duplicated commit
The problem is that the range to the previous branch includes not just the single commit on the current feature branch, but also the commits from the previous feature branch that have now been rebased (on top of the integrated previous-previous branch).
Git doesn't know how to rebase those, and we also cannot manually provide references, as the submodule has been merged as well (and even if it'd be very cumbersome).
If we just add the conflicted submodules, we'll wrongly incorporate the current (updated) submodule state into a previous superproject commit, and that single updated commit we're actually after will be gone!
The solution lies in recognizing that we're just interested in updating the last commit in the superproject, so we have to tell Git to discard any previous ones, knowing that those already exist (in rebased form, and therefore inaccessible to Git) in the previous branch we're rebasing onto.
There's still the conflict, but this time we can just add the current submodule state, and that'll be the new commit we're after.